+++ libroad/hwm_texture_gen.cpp
@@ -203,8 +203,8 @@ namespace hwm
         if(max_h == 0.0)
             max_h = 1.0;
 
-        vec2u minsize(static_cast<unsigned int>(std::ceil(max_h/min_y_feat)),
-                      static_cast<unsigned int>(std::ceil(total_w/min_x_feat)));
+        vec2u minsize(static_cast<size_t>(std::ceil(max_h/min_y_feat)),
+                      static_cast<size_t>(std::ceil(total_w/min_x_feat)));
         im_res = 3*minsize;
         for(int i = 0; i < 2; ++i)
             im_res[i] = std::max(im_res[i], (size_t)(1));
