+++ test/displace-polylines.cpp
@@ -22,7 +22,7 @@ int main(int argc, char *argv[])
 
     im.flip();
 
-    vec2i dim(im.columns(), im.rows());
+    vec2i dim((int)im.columns(), (int)im.rows());
 
     float *pix = new float[dim[0]*dim[1]];
     im.write(0, 0, dim[1], dim[0], "R", Magick::FloatPixel, pix);
@@ -52,7 +52,7 @@ int main(int argc, char *argv[])
     arc_road ar;
     ar.initialize_from_polyline(0.0f, poly);
     size_t reverse_start;
-    ar.make_mesh(vrts, fcs, reverse_start, vec2f(0, 1), vec2f(-0.1, 0.1), 0.01);
+    ar.make_mesh(vrts, fcs, reverse_start, vec2f(0.0f, 1.0f), vec2f(-0.1f, 0.1f), 0.01);
 
     mesh_to_obj(os, "poly", "poly", vrts, fcs);
